Split "welcome" actions/reducers into "conference" and "connection"
