+++ features/conference/actionTypes.js
@@ -1,0 +1,13 @@
+/**
+ * Action type to signal that we are joining the conference.
+ *
+ * @type {string}
+ */
+export const JITSI_CONFERENCE_JOINED = 'JITSI_CONFERENCE_JOINED';
+
+/**
+ * Action type to signal that we are leaving the conference.
+ *
+ * @type {string}
+ */
+export const JITSI_CONFERENCE_LEFT = 'JITSI_CONFERENCE_LEFT';
