+++ features/conference/index.js
@@ -1,1 +1,2 @@
+export * from './actions';
 export * from './components';
