+++ features/welcome/index.js
@@ -1,2 +1,1 @@
-export * from './actions';
 export * from './components';
